Removed MainActivity + unused resources
