+++ app/src/main/AndroidManifest.xml
@@ -24,13 +24,8 @@
         <activity
             android:name=".activity.CategoriesActivity"
             android:label="@string/app_name"
-            android:parentActivityName=".activity.MainActivity"
             android:windowSoftInputMode="stateHidden">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.example.enigmator.activity.MainActivity" />
         </activity>
-        <activity android:name=".activity.MainActivity" />
         <activity
             android:name=".activity.TopicActivity"
             android:parentActivityName=".activity.CategoriesActivity"/>
