+++ app/src/main/res/drawable/ic_games_purple_24dp.xml
