--- app/src/main/res/layout/statistic_item.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent" android:layout_height="match_parent"
-    android:orientation="horizontal"
-    android:padding="@dimen/list_item_padding">
-
-    <TextView
-        android:id="@+id/text_stat_title"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:textSize="@dimen/list_item_text_size"
-        android:textColor="@color/colorPrimaryDark"
-        android:layout_weight="0.4"/>
-
-    <TextView
-        android:id="@+id/text_stat_content"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:textSize="@dimen/list_item_text_size"
-        android:textColor="@color/black"
-        android:layout_weight="0.6"/>
-
-</LinearLayout>
