+++ app/src/main/res/values/dimens.xml
@@ -10,6 +10,5 @@
 
     <dimen name="list_item_padding">8dp</dimen>
 
-    <dimen name="design_bottom_navigation_text_size" tools:override="true">12sp</dimen>
     <dimen name="design_bottom_navigation_active_text_size" tools:override="true">12sp</dimen>
 </resources>
