+++ build.gradle
@@ -7,7 +7,7 @@ buildscript {
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.0'
+        classpath 'com.android.tools.build:gradle:3.4.1'
         
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
